rename fabric file
