+++ tests/tests_lite/test_parity.py
@@ -30,7 +30,7 @@ from torch.nn.parallel.distributed import DistributedDataParallel
 from torch.utils.data import DataLoader
 from torch.utils.data.distributed import DistributedSampler
 
-from lightning_lite.lite import Fabric
+from lightning_lite.fabric import Fabric
 from lightning_lite.plugins.environments.lightning import find_free_network_port
 from lightning_lite.strategies.ddp_spawn import DDPSpawnStrategy
 from lightning_lite.utilities.apply_func import move_data_to_device
